libs/genesis-simulator/pyproject.toml
[project]
name = "genesis-simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "genesis-world>=0.2.0",
    "torch>=2.5.1",
]
